requirements.txt
google-genai==1.2.0
google-generativeai==0.8.4
dotenv
langchain_community
langgraph 
langchain_anthropic